+++ examples/selectHelper.js
@@ -1,0 +1,4 @@
+import {augmentSelect} from "../source/virstellung.js";
+
+augmentSelect("imageS");
+augmentSelect("imageS2");