+++ examples/tools/rollup.config.js
@@ -4,31 +4,12 @@ import resolve from '@rollup/plugin-node-resolve';
 const GLOBAL_NAME = `x`;
 
 const commonOutputOptions = {
-    // core output options
     name: GLOBAL_NAME,
-    // globals: [],
-
-    // advanced output options
-    // paths: {},
-    // footer: ``,
-    // intro: ``,
-    // outro: ``,
-    // sourcemap,
-    // sourcemapFile,
     interop: false,
     extend: false,
 
-    // danger zone
-    // exports,
-    // indent,
     strict: true,
-    // freeze,
     namespaceToStringTag: false
-
-    // experimental
-    // entryFileNames,
-    // chunkFileNames,
-    // assetFileNames
 };
 
 export default [{
@@ -42,7 +23,25 @@ export default [{
     output: [
         Object.assign({
             format: `es`,
-            file: `built/virstellung.es.js`,
+            file: `built/virstellungAutoLaunch.es.js`,
+        }, commonOutputOptions),
+    ],
+
+    watch: {
+        clearScreen: true
+    }
+},{
+    input: `./selectHelper.js`,
+    plugins: [resolve()],
+    treeshake: {
+        moduleSideEffects: true,
+        moduleSideEffects: `no-external`,
+    },
+
+    output: [
+        Object.assign({
+            format: `es`,
+            file: `built/selectHelper.es.js`,
         }, commonOutputOptions),
     ],